pyproject.toml
[tool.poetry]
name = "hackernews-podcasts"
version = "0.1.0"
description = "A simple HackerNews article podcastfy implementation"
authors = ["madhavarora1988@gmail.com"]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.28.0"
python-dotenv = "^0.20.0"
schedule = "^1.1.0"
pandas = "^2.0.0"
openpyxl = "^3.1.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
black = "^22.3.0"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
